requirements.txt
python-dotenv
planetary-computer
pystac-client
Pillow
requests
